adding further billing tests

+++ src/main/java/uk/thepragmaticdev/billing/dto/response/BillingPriceRecurringResponse.java
@@ -11,5 +11,5 @@ public class BillingPriceRecurringResponse {
 
   private String interval;
 
-  private int intervalCount;
+  private long intervalCount;
 }
