+++ src/main/java/uk/thepragmaticdev/billing/dto/response/BillingPriceTierResponse.java
@@ -1,5 +1,6 @@
 package uk.thepragmaticdev.billing.dto.response;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,7 +12,7 @@ public class BillingPriceTierResponse {
 
   private long flatAmount;
 
-  private float unitAmountDecimal;
+  private BigDecimal unitAmountDecimal;
 
   private long upTo;
 }
